refactor(download): move parse_timestamp tests beside the util and document strict parsing
